requirements.txt
attrs==23.2.0
aws-cdk-lib==2.55.1
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.166
black==24.1.0
cattrs==23.2.3
click==8.1.7
constructs==10.3.0
importlib-resources==6.1.1
invoke==2.2.0
jsii==1.94.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.9.0
